# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Service for managing event schedules with risk analysis and optimization",
        "title": "Event Scheduler API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8282",
    "basePath": "/api/v1",
    "paths": {
...
        "/schedules/{id}/public": {
            "get": {
                "description": "Get a formatted public version of a schedule",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedules"
                ],
                "summary": "Get public schedule",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Schedule ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "json",
                            "text"
                        ],
                        "type": "string",
                        "default": "json",
                        "description": "Output format (json or text)",
                        "name": "format",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "When format=text",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/schedules/{id}/volunteer": {
            "get": {
                "description": "Get a formatted version of a schedule for volunteers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedules"
                ],
                "summary": "Get volunteer schedule",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Schedule ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cor-events-scheduler_internal_domain_models.VolunteerSchedule"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "cor-events-scheduler_internal_domain_models.VolunteerSchedule": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string",
                    "example": "2024-07-01T00:00:00Z"
                },
                "event_name": {
                    "type": "string",
                    "example": "Summer Music Festival"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cor-events-scheduler_internal_domain_models.VolunteerScheduleItem"
                    }
                },
                "notes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"Wear volunteer badge\"",
                        "\"Follow safety guidelines\"]"
                    ]
                }
            }
        },
        "cor-events-scheduler_internal_domain_models.VolunteerScheduleItem": {
            "type": "object",
            "properties": {
                "break_duration": {
                    "type": "integer",
                    "example": 15
                },
                "equipment": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"Microphones\"",
                        "\"Speakers\"]"
                    ]
                },
                "instructions": {
                    "type": "string",
                    "example": "Help with equipment setup"
                },
                "location": {
                    "type": "string",
                    "example": "Main Stage"
                },
                "required_staff": {
                    "type": "integer",
                    "example": 5
                },
                "setup_notes": {
                    "type": "string",
                    "example": "Check all connections"
                },
                "tech_break": {
                    "type": "boolean",
                    "example": false
                },
                "time": {
                    "type": "string",
                    "example": "2024-07-01T13:00:00Z"
                },
                "title": {
                    "type": "string",
                    "example": "Stage Setup"
                }
            }
        },
...
        "internal_handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
    }
}